+++ push_swap/libft/srcs/strings/ft_atoi.c
@@ -6,7 +6,7 @@
 /*   By: gjacome- <gjacome-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/08 15:23:47 by gjacome-          #+#    #+#             */
-/*   Updated: 2024/08/22 12:21:54 by gjacome-         ###   ########.fr       */
+/*   Updated: 2025/03/20 17:35:55 by gjacome-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,11 +22,11 @@ static int	ft_isspace(char c)
 		|| c == ' ');
 }
 
-int	ft_atoi(const char *str)
+long	ft_atoi(const char *str)
 {
 	int	i;
 	int	sign;
-	int	total;
+	long	total;
 
 	sign = 1;
 	i = 0;
